refactor: create external exist function and move throwing custom errors to try block
