+++ src/fs/create.js
@@ -1,7 +1,8 @@
 import { access, appendFile } from 'fs/promises';
 import { fileURLToPath } from 'url';
 import { dirname, join } from 'path';
-import CustomFSError from './CustomError.js';
+import CustomFSError from './utils/CustomError.js';
+import { isExists } from './utils/exist.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -12,15 +13,15 @@ export const create = async () => {
   const content = 'I am fresh and young';
 
   try {
-    await access(pathToFile);
-    throw new CustomFSError('FS operation failed');
+    const isFileExists = await isExists(pathToFile);
+    if (isFileExists) {
+      throw new CustomFSError('FS operation failed');
+    }
+    await appendFile(pathToFile, content);
+    console.log('File was successfuly created') 
   }
   catch(err) {
-    if (err instanceof CustomFSError) { console.log(err); }
-    else { 
-      await appendFile(pathToFile, content);
-      console.log('File was successfuly created') 
-    }
+    console.log(err);
   }
 
 };
